Update navigation and view components; add bootstrap and vue-sweetalert2
The HelloWorld component was removed and new components related to employees were added to manage different actions (create, list, edit). Bootstrap was implemented to improve the UX and UI in the App.vue file offering a better navigation bar. A third-party library, vue-sweetalert2 was also included in the package dependencies to handle alerts in the application.